requirements.txt
apex==0.1
matplotlib==3.5.2
numpy==1.21.5
opencv_python==4.6.0.66
pandas==1.3.5
pysodmetrics==1.3.1
PyYAML==6.0
signatory==1.2.6.1.8.0
tensorboardX==2.5.1
thop==0.1.1.post2209072238
torch==1.8.0
tqdm==4.64.0
